+++ src/main/java/com/trego/dto/BucketDTO.java
@@ -2,6 +2,8 @@ package com.trego.dto;
 
 import lombok.Data;
 import com.fasterxml.jackson.annotation.JsonInclude;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -12,10 +14,10 @@ public class BucketDTO {
     private List<BucketItemDTO> availableItems; // Available medicines
     private List<UnavailableMedicineDTO> unavailableItems; // Unavailable medicines with substitutes
     private List<SelectedSubstituteDTO> selectedSubstitutes; // User-selected substitutes
-    private double totalPrice;
-    private double deliveryCharges;
-    private double amountToPay;
-    private double totalDiscount; // Total discount across all items in the bucket
+    private BigDecimal totalPrice;
+    private BigDecimal deliveryCharges;
+    private BigDecimal amountToPay;
+    private BigDecimal totalDiscount; // Total discount across all items in the bucket
     private Long vendorId; // If all items are from the same vendor
     private String vendorName; // If all items are from the same vendor
     private String logo;
