+++ src/main/java/com/trego/dto/response/PreOrderResponseDTO.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.trego.dto.CartDTO;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -13,10 +14,10 @@ public class PreOrderResponseDTO {
     private long userId;
     private long orderId;
     private long addressId;
-    private double totalCartValue;
-    private double amountToPay;
-    private double discount;
-    private double deliveryCharges;
+    private BigDecimal totalCartValue;
+    private BigDecimal amountToPay;
+    private BigDecimal discount;
+    private BigDecimal deliveryCharges;
 
     private List<CartResponseDTO> carts;
